+++ tests/Config/ConfigFile/ConfigFileTest.php
@@ -11,6 +11,7 @@
 
 namespace Puli\RepositoryManager\Tests\Config\ConfigFile;
 
+use Puli\RepositoryManager\Config\Config;
 use Puli\RepositoryManager\Config\ConfigFile\ConfigFile;
 
 /**
@@ -52,5 +53,19 @@ class ConfigFileTest extends \PHPUnit_Framework_TestCase
     public function testPathMustBeValid($invalidPath)
     {
         new ConfigFile($invalidPath);
+    }
+
+    public function testBaseConfig()
+    {
+        $baseConfig = new Config();
+        $configFile = new ConfigFile('/path', $baseConfig);
+        $config = $configFile->getConfig();
+
+        $this->assertNotSame($baseConfig, $config);
+
+        $baseConfig->set(Config::PULI_DIR, 'my-puli-dir');
+
+        $this->assertSame('my-puli-dir', $config->get(Config::PULI_DIR));
+        $this->assertNull($config->get(Config::PULI_DIR, false));
     }
 }